requirements.txt
streamlit>=1.33
opencv-python-headless>=4.9.0.80
numpy>=1.23
Pillow>=10.0.0
